allo_justice_django/requirements.txt
appdirs
asgiref
astroid
autopep8
bandit
certifi
cffi
chardet
colorama
cryptography
distlib
Django
django-clear-cache
django-crispy-forms
django-dotenv
django-embed-video
django-session-timeout
django-six
django-utils-six
dodgy
filelock
flake8
flake8-polyfill
gitdb
GitPython
idna
isort
lazy-object-proxy
livereload
mccabe
mypy
mypy-extensions
numpy
pandas
pbr
pep8-naming
Pillow
prospector
psycopg2
pycodestyle
pycparser
pydocstyle
pyflakes
pylint
pylint-celery
pylint-django
pylint-flask
pylint-plugin-utils
python-dateutil
pytz
PyYAML
requests
requirements-detector
setoptconf
six
smmap
snowballstemmer
sqlparse
stevedore
toml
tornado
typed-ast
typing-extensions
urllib3
virtualenv
whitenoise
wrapt
xlwt
django-allauth
